Remove decreasing results per page logic
Doing this is a little more complex than first imagined. The issue is
that when you lower the results per page, the "current" page you are
on changes. For example, if you're on page 2 with 100 results per
page, changing to 50 results per page puts you on page 4. And it
only gets more complicated